+++ src/org/myrobotlab/service/Tracking.java
@@ -111,8 +111,8 @@ public class Tracking extends Service {
 	// MRL points
 	public Point2Df lastPoint = new Point2Df();
 
-	private Double lastXServoPos;
-	private Double lastYServoPos;
+	private Integer lastXServoPos;
+	private Integer lastYServoPos;
 
 	// ----- INITIALIZATION DATA BEGIN -----
 	public double xSetpoint = 0.5;
@@ -437,7 +437,7 @@ public class Tracking extends Service {
 				faceFoundFrameCount = 0;
 
 				if (scan) {
-					double xpos = x.getPos();
+					int xpos = x.getPos();
 
 					if (xpos + scanXStep >= x.getMax() && scanXStep > 0 || xpos + scanXStep <= x.getMin() && scanXStep < 0) {
 						scanXStep = scanXStep * -1;
@@ -459,7 +459,7 @@ public class Tracking extends Service {
 
 		// FIXME - remove not used
 		case STATE_FACE_DETECT_LOST_TRACK:
-			double xpos = x.getPos();
+			int xpos = x.getPos();
 
 			if (xpos >= x.getMax() && scanXStep > 0) {
 				scanXStep = scanXStep * -1;
@@ -594,8 +594,8 @@ public class Tracking extends Service {
 
 		pid.setInput("x", targetPoint.x);
 		pid.setInput("y", targetPoint.y);
-		double currentXServoPos = x.getPos();
-		double currentYServoPos = y.getPos();
+		int currentXServoPos = x.getPos();
+		int currentYServoPos = y.getPos();
 
 		// TODO - work on removing currentX/YServoPos - and use the servo's
 		// directly ???
